Move getMaxRetries to WebhookInterface implementation from WebhookRegistry
